refactor: reorganize global actions code
