+++ src/pages/global_actions/page.js
@@ -1,32 +1,31 @@
-import { getCurrentProject, getCurrentProjectEditor } from '../app/main.js';
-import { decToHex } from '../common/character_ids.js';
-import { addAsChildren, makeElement, textToNode } from '../common/dom.js';
-import { remove } from '../common/functions.js';
+import { getCurrentProject, getCurrentProjectEditor } from '../../app/main.js';
+import { decToHex } from '../../common/character_ids.js';
+import { addAsChildren, makeElement, textToNode } from '../../common/dom.js';
+import { remove } from '../../common/functions.js';
 import {
 	closeAllModalDialogs,
 	showError,
 	showModalDialog,
 	showToast,
-} from '../controls/dialogs/dialogs.js';
-import { CharacterRange } from '../project_data/character_range.js';
-import { glyphChanged } from '../project_editor/cross_item_actions.js';
-import { makeNavButton, toggleNavDropdown } from '../project_editor/navigator.js';
-import {
-	makeCard_AllCaps,
-	makeCard_Diacritics,
-	makeCard_DiacriticsAdvanced,
-	makeCard_Flatten,
-	makeCard_Monospace,
-	makeCard_Move,
-	makeCard_RemoveItems,
-	makeCard_Resize,
-	makeCard_Round,
-	makeCard_ScaleHorizontal,
-	makeCard_ScaleVertical,
-	makeCard_SideBearings,
-	makeCard_Skew,
-} from './global_actions_cards.js';
-import { updateAllCharacterRangeCounts } from './settings_project.js';
+} from '../../controls/dialogs/dialogs.js';
+import { CharacterRange } from '../../project_data/character_range.js';
+import { glyphChanged } from '../../project_editor/cross_item_actions.js';
+import { makeNavButton, toggleNavDropdown } from '../../project_editor/navigator.js';
+import { makeCard_Diacritics } from './diacritics.js';
+import { makeCard_DiacriticsAdvanced } from './diacritics.js';
+import { makeCard_AllCaps } from './font_types.js';
+import { makeCard_Flatten } from './project_cleanup.js';
+import { makeCard_Monospace } from './font_types.js';
+import { makeCard_Move } from './move_and_resize.js';
+import { makeCard_RemoveItems } from './project_cleanup.js';
+import { makeCard_Resize } from './move_and_resize.js';
+import { makeCard_Round } from './project_cleanup.js';
+import { makeCard_ScaleHorizontal } from './move_and_resize.js';
+import { makeCard_ScaleVertical } from './move_and_resize.js';
+import { makeCard_SideBearings } from './move_and_resize.js';
+import { makeCard_Skew } from './move_and_resize.js';
+
+import { updateAllCharacterRangeCounts } from '../settings_project.js';
 /**
  * Page > Global Actions
  * Various actions that can be applied to all glyphs.
